Move layout struct to engines
