+++ pkg/engines/layout.go
@@ -1,4 +1,4 @@
-package utils
+package engines
 
 type Layout struct {
 	Symlink string `yaml:"symlink"`
