+++ pkg/models/types.go
@@ -1,6 +1,8 @@
 package models
 
-import "github.com/canonical/inference-snaps-cli/v2/pkg/utils"
+import (
+	"github.com/canonical/inference-snaps-cli/v2/pkg/engines"
+)
 
 type Manifest struct {
 	ID   string `json:"id" yaml:"id"`
@@ -17,5 +19,5 @@ type Manifest struct {
 
 	Environment []string `json:"environment" yaml:"environment"`
 
-	Layout map[string]utils.Layout `json:"layout,omitempty" yaml:"layout,omitempty"`
+	Layout map[string]engines.Layout `json:"layout,omitempty" yaml:"layout,omitempty"`
 }
