+++ pkg/runtimes/types.go
@@ -1,13 +1,15 @@
 package runtimes
 
-import "github.com/canonical/inference-snaps-cli/v2/pkg/utils"
+import (
+	"github.com/canonical/inference-snaps-cli/v2/pkg/engines"
+)
 
 type Manifest struct {
-	Name        string                  `json:"name" yaml:"name"`
-	Servers     map[string]Server       `json:"servers" yaml:"servers"`
-	Environment []string                `json:"environment" yaml:"environment"`
-	Layout      map[string]utils.Layout `yaml:"layout"`
-	Components  []string                `json:"components" yaml:"components"`
+	Name        string                    `json:"name" yaml:"name"`
+	Servers     map[string]Server         `json:"servers" yaml:"servers"`
+	Environment []string                  `json:"environment" yaml:"environment"`
+	Layout      map[string]engines.Layout `yaml:"layout"`
+	Components  []string                  `json:"components" yaml:"components"`
 }
 
 type Server struct {
